搜索界面布局完成

+++ spade/spade/Classes/Model/SPDashBoard.swift
@@ -12,7 +12,7 @@ import YYModel
 class SPDashBoard: NSObject {
     
     /// id
-    var id: Int = 0
+    var id: Int64 = 0
     /// 点赞要用
     var reblog_key: String?
     // 内容
